Remove batch stubs from DevKit install script.
Originally, this behavior was implemented as a fallback safety net
if a RubyGems installation was not found. This behavior has been
removed as all current RubyInstaller installations contain RubyGems
and these batch stubs create potential system conflicts. If this
"global" PATH behavior is desired, use the devkitvars.bat or
devkitvars.ps1 helper scripts to bring the DevKit artifacts onto
PATH.

The new behavior is to warn the user and exit the install script
if a RubyGems installation is not found.