Add a YAML syntax hint to dk.rb
Since config.yml needs to be valid YAML, simply adding a line with the path to ruby at the end of it will cause the file to be invalid, resulting in an error when dk.rb install is run. So, include a hit about YAML syntax in the instructions at the top of the config.yml file.